+++ engine/src/Core/Interfaces/IParticleWobble.ts
@@ -1,5 +1,0 @@
-export interface IParticleWobble {
-    angle: number;
-    angleSpeed: number;
-    moveSpeed: number;
-}