+++ plugins/sounds/src/Options/Classes/SoundsIcon.ts
@@ -2,8 +2,15 @@ import type { IOptionLoader, RecursivePartial } from "tsparticles-engine";
 import type { ISoundsIcon } from "../Interfaces/ISoundsIcon";
 
 export class SoundsIcon implements ISoundsIcon, IOptionLoader<ISoundsIcon> {
+    height;
     path?: string;
     svg?: string;
+    width;
+
+    constructor() {
+        this.width = 24;
+        this.height = 24;
+    }
 
     load(data?: RecursivePartial<ISoundsIcon>): void {
         if (!data) {
@@ -16,6 +23,14 @@ export class SoundsIcon implements ISoundsIcon, IOptionLoader<ISoundsIcon> {
 
         if (data.svg !== undefined) {
             this.svg = data.svg;
+        }
+
+        if (data.width !== undefined) {
+            this.width = data.width;
+        }
+
+        if (data.height !== undefined) {
+            this.height = data.height;
         }
     }
 }