+++ plugins/sounds/src/Options/Interfaces/ISoundsEvent.ts
@@ -1,10 +1,11 @@
 import type { FilterFunction } from "../../types";
+import type { ISoundsAudio } from "./ISoundsAudio";
 import type { ISoundsMelody } from "./ISoundsMelody";
 import type { ISoundsNote } from "./ISoundsNote";
 import type { SingleOrMultiple } from "tsparticles-engine";
 
 export interface ISoundsEvent {
-    audio?: SingleOrMultiple<string>;
+    audio?: SingleOrMultiple<ISoundsAudio | string>;
 
     event: SingleOrMultiple<string>;