+++ plugins/sounds/src/Options/Interfaces/ISoundsVolume.ts
@@ -1,0 +1,6 @@
+export interface ISoundsVolume {
+    max: number;
+    min: number;
+    step: number;
+    value: number;
+}