+++ plugins/sounds/src/enums.ts
@@ -1,0 +1,4 @@
+export const enum SoundsEventType {
+    mute = "soundsMuted",
+    unmute = "soundsUnmuted",
+}