+++ components/angular/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.1](https://github.com/matteobruni/tsparticles/compare/ng-particles@3.7.0...ng-particles@3.7.1) (2022-12-25)
+
+**Note:** Version bump only for package ng-particles
+
 # [3.7.0](https://github.com/matteobruni/tsparticles/compare/ng-particles@3.6.0...ng-particles@3.7.0) (2022-12-23)
 
 **Note:** Version bump only for package ng-particles