+++ components/vue-alt/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.1](https://github.com/matteobruni/tsparticles/compare/particles.vue@3.7.0...particles.vue@3.7.1) (2022-12-25)
+
+**Note:** Version bump only for package particles.vue
+
 # [3.7.0](https://github.com/matteobruni/tsparticles/compare/particles.vue@3.6.0...particles.vue@3.7.0) (2022-12-23)
 
 **Note:** Version bump only for package particles.vue